refactor!: Move `_main_audio_processor` from `player.py` to `main_audio_processor` to `pipes.py`
Technical Debt - Decouple main audio processor into different generators with more explicit argument passthrough
